+++ pywb/warc/pathresolvers.py
@@ -59,18 +59,15 @@ class RedisResolver:
 class PathIndexResolver:
     def __init__(self, pathindex_file):
         self.pathindex_file = pathindex_file
-        self.reader = open(pathindex_file, 'rb')
 
     def __call__(self, filename):
-        result = iter_exact(self.reader, filename, '\t')
+        with open(self.pathindex_file, 'rb') as reader:
+            result = iter_exact(reader, filename, '\t')
 
-        def gen_list(result):
             for pathline in result:
                 paths = pathline.split('\t')[1:]
                 for path in paths:
                     yield path
-
-        return gen_list(result)
 
     def __repr__(self):  # pragma: no cover
         return "PathIndexResolver('{0}')".format(self.pathindex_file)